Move tmpdir to legacypath plugin
