+++ doc/en/reference/reference.rst
@@ -638,7 +638,7 @@ tmpdir
 
 :ref:`tmpdir and tmpdir_factory`
 
-.. autofunction:: _pytest.tmpdir.tmpdir()
+.. autofunction:: _pytest.legacypath.tmpdir()
     :no-auto-options:
